Refactoring skills category section
