+++ src/app/character-skills/page.tsx
@@ -1,38 +1,14 @@
 
-"use client"
-import React, { useState } from "react";
-import Link from 'next/link';
-import Layout from '../../shared/components/Layout';
-import Categories from '../../shared/components/Skill_categories';
-import SearchBar from '../../shared/components/SearchBar/SearchBar';
-
+"use client";
+import Link from "next/link";
+import Layout from "../../shared/components/Layout";
+import SkillsCategorySection from "./components/SkillsCategorySection";
 
 function CharacterSkill() {
-  const [searchQuery, setSearchQuery] = useState('');
-
-  const handleSearchChange = (value: string) => {
-    setSearchQuery(value);
-  };
-
-  const handleSearchSubmit = () => {
-    console.log("Submit: Searching for:", searchQuery);
-    
-  };
-
   return (
     <Layout>
       <h1>Character Skill</h1>
-
-      {}
-      <SearchBar
-        searchQuery={searchQuery}
-        onSearchChange={handleSearchChange}
-        onSearchSubmit={handleSearchSubmit}
-      />
-
-      
-      <Categories />
-
+      <SkillsCategorySection />
       <Link href="/backstory">Przejdź do historii postaci</Link>
     </Layout>
   );
